break out loading of client data in a separate file
