+++ web/src/routes/dashboard/+page.svelte
@@ -1,53 +1,16 @@
 <script lang="ts">
   import { getSelectedWorkspace, getSelectedExperiment } from "./state.svelte";
-  import type {
-    PendingInvitation,
-    ApiResponse,
-    WorkspaceRole,
-  } from "$lib/types";
-  import { onMount } from "svelte";
   import WorkspaceColumn from "./WorkspaceColumn.svelte";
   import ExperimentListColumn from "./ExperimentListColumn.svelte";
   import ExperimentDetails from "./ExperimentDetails.svelte";
   import EmptyState from "./EmptyState.svelte";
+  import { createWorkspaceData } from "./workspaceData.svelte";
 
   let { data } = $props();
   let workspaces = $derived(data.workspaces);
-  let workspaceRoles = $state<WorkspaceRole[]>([]);
-  let workspaceInvitations = $state<PendingInvitation[]>([]);
   let selectedWorkspace = $derived(getSelectedWorkspace());
   let selectedExperiment = $derived(getSelectedExperiment());
-
-  async function loadWorkspaceRoles() {
-    try {
-      const response = await fetch("/api/workspace-roles");
-      if (!response.ok) {
-        throw new Error(`HTTP ${response.status}: ${response.statusText}`);
-      }
-      const apiReponse: ApiResponse<WorkspaceRole[]> = await response.json();
-      workspaceRoles = apiReponse.data.filter((item) => {
-        return item.name !== "OWNER";
-      });
-    } catch (error) {
-      console.error(`Failed to load workspace roles:`, error);
-    }
-  }
-
-  async function loadPendingInvitations() {
-    try {
-      const response = await fetch("/api/workspace-invitations");
-      const responseJson: ApiResponse<PendingInvitation[]> =
-        await response.json();
-      workspaceInvitations = responseJson.data;
-    } catch (error) {
-      console.error("Failed to load pending invitations:", error);
-    }
-  }
-
-  onMount(async () => {
-    await loadWorkspaceRoles();
-    await loadPendingInvitations();
-  });
+  const { workspaceRoles, workspaceInvitations } = createWorkspaceData();
 </script>
 
 <div
