+++ web/src/routes/dashboard/workspaceData.svelte.ts
@@ -1,0 +1,49 @@
+import { onMount } from "svelte";
+import type { ApiResponse, WorkspaceRole, PendingInvitation } from "$lib/types";
+
+export function createWorkspaceData() {
+  let workspaceRoles = $state<WorkspaceRole[]>([]);
+  let workspaceInvitations = $state<PendingInvitation[]>([]);
+
+  async function loadWorkspaceRoles() {
+    try {
+      const response = await fetch("/api/workspace-roles");
+      if (!response.ok) {
+        throw new Error(`HTTP ${response.status}: ${response.statusText}`);
+      }
+      const apiReponse: ApiResponse<WorkspaceRole[]> = await response.json();
+      workspaceRoles = apiReponse.data.filter((item) => {
+        return item.name !== "OWNER";
+      });
+    } catch (error) {
+      console.error(`Failed to load workspace roles:`, error);
+    }
+  }
+
+  async function loadPendingInvitations() {
+    try {
+      const response = await fetch("/api/workspace-invitations");
+      const responseJson: ApiResponse<PendingInvitation[]> =
+        await response.json();
+      workspaceInvitations = responseJson.data;
+    } catch (error) {
+      console.error("Failed to load pending invitations:", error);
+    }
+  }
+
+  onMount(async () => {
+    await loadWorkspaceRoles();
+    await loadPendingInvitations();
+  });
+
+  return {
+    // get workspaceRoles() {
+    //   return workspaceRoles;
+    // },
+    // get workspaceInvitations() {
+    //   return workspaceInvitations;
+    // },
+    workspaceRoles,
+    workspaceInvitations,
+  };
+}
